requirements.txt
requests==2.25.0
pyyaml==5.3.1
aws-psycopg2
boto3==1.17.23
